Cleaner "Offer a page reload for users" recipe
Since Workbox v6 a new method, messageSkipWaiting(), has been added to the workbox-window module to simplify the process of telling the "waiting" service worker to activate. And all "external" events in workbox-window have been removed in place of "normal" events with an isExternal property set to true. Thanks to both of the above changes, the "Offer a page reload for users" recipe can be simplified.